+++ demo/src/main/java/com/example/demo/dto/incoming/DTO_CreateOrder.java
@@ -3,13 +3,13 @@ package com.example.demo.dto.incoming;
 import java.util.List;
 
 public class DTO_CreateOrder {
-    private int customerID;
+    private long customerID;
     private List<DTO_CartProduct> productQuantity;
 
-    public int getCustomerID() {
+    public long getCustomerID() {
         return customerID;
     }
-    public void setCustomerID(int customerID) {
+    public void setCustomerID(long customerID) {
         this.customerID = customerID;
     }
     public List<DTO_CartProduct> getProductQuantity() {
